backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "FastAPI backend for Todo Full-Stack Web Application"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "sqlmodel==0.0.14",
    "asyncpg==0.29.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
